+++ iml-wire-types/Cargo.toml
@@ -11,6 +11,7 @@ bytes = {version = "0.5", optional = true}
 chrono = {version = "0.4", features = ["wasmbind", "serde"]}
 humantime = "2.0"
 im = {version = "15.0", features = ["serde"]}
+iml-api-utils = {path = "../iml-api-utils", version = "0.3"}
 iml-change = {path = "../iml-change", version = "0.1", optional = true}
 juniper = {git = "https://github.com/graphql-rust/juniper", optional = true}
 lazy_static = "1.4.0"
